requirements.txt
# Core Telegram & OpenAI
telethon>=1.40.0
openai>=1.0.0

# Streamlit Web Interface
streamlit>=1.28.0

# Vector Database & RAG
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Web Scraping & Crawling
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
html5lib>=1.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Crawler Dependencies
sqlite3  # Built-in with Python
concurrent.futures  # Built-in with Python
threading  # Built-in with Python
queue  # Built-in with Python
hashlib  # Built-in with Python
urllib.robotparser  # Built-in with Python
urllib.parse  # Built-in with Python

# Optional Performance Packages
cryptg>=0.4.0
pillow>=10.0.0
aiohttp>=3.9.0
hachoir>=3.2.0

# Additional Utilities
python-dateutil>=2.8.0
urllib3>=2.0.0
pytz>=2023.3

# Environment Variables
python-dotenv>=1.0.0 